SLOPPY COMMIT
This is a scratch pad for playing with turbo streams and stimulus JS

The core chunk is the Part which has some validations we want to display
when the model isn't valid anymore. Right now you hit submit to persist
changes but you could change the event handler on the form to submit on
focus exit instead and get a real time validation engine which syncs
back to the model.

I'm not sure how we could create the validations for our data and have
them sync back to the form while still saving the underlying data. It
could be that introducing a form object could let us have an invalid
form while still persisting the data.